+++ server/index.js
@@ -7,7 +7,8 @@ const cookieParser = require("cookie-parser");
 const UserModel = require("./models/User");
 const varifyUser = require("./middlewares/verifyUser.js");
 const dotenv = require("dotenv");
-const inputValidation = require("./middlewares/inputValidation.js");
+
+const userRouter = require("./routes/users.js");
 
 const app = express();
 app.use(express.json());
@@ -26,41 +27,7 @@ app.get("/dashboard", varifyUser, (req, res) => {
   res.json("Success");
 });
 
-// can add ZOD for input validation
-app.post("/register", inputValidation, (req, res) => {
-  const { name, email, password } = req.body;
-  bcrypt
-    .hash(password, 10)
-    .then((hash) => {
-      UserModel.create({ name, email, password: hash })
-        .then((user) => res.json("Success"))
-        .catch((err) => res.json(err));
-    })
-    .catch((err) => res.json(err));
-});
-
-app.post("/login", inputValidation, (req, res) => {
-  const { email, password } = req.body;
-  UserModel.findOne({ email: email }).then((user) => {
-    if (user) {
-      bcrypt.compare(password, user.password, (err, response) => {
-        if (response) {
-          const token = jwt.sign(
-            { email: user.email, role: user.role },
-            "jwt-secret-key",
-            { expiresIn: "1d" }
-          );
-          res.cookie("token", token);
-          return res.json({ Status: "Success", role: user.role });
-        } else {
-          return res.json("The password is incorrect");
-        }
-      });
-    } else {
-      return res.json("No record existed");
-    }
-  });
-});
+app.use("/user", userRouter);
 
 app.use(function (req, res, next) {
   res.header("Access-Control-Allow-Origin", "http://127.0.0.1:5173"); // Replace with your client-side origin
